requirements.txt
alembic==1.5.4
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
cymem==2.0.5
elasticsearch==7.11.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pathy==0.4.0
pdfminer.six==20201018
preshed==3.0.5
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
spacy==3.0.3
spacy-legacy==3.0.1
SQLAlchemy==1.3.23
srsly==2.4.0
thinc==8.0.1
tqdm==4.57.0
typer==0.3.2
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
WTForms==2.3.3
